+++ index.html
@@ -113,9 +113,11 @@
   <body>
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
-    <!--meta pixel-->
+    <!--
+    meta pixel
     <script>
       fbq('track', 'Purchase', {currency: "USD", value: 30.00});
     </script>
+    -->
   </body>
 </html>
