+++ package.json
@@ -14,8 +14,10 @@
     "react": "^18.3.1",
     "react-copy-to-clipboard": "^5.1.0",
     "react-dom": "^18.3.1",
+    "react-lite-youtube-embed": "^2.4.0",
     "react-markdown": "^9.0.1",
     "react-router-dom": "^6.26.2",
+    "react-youtube": "^10.1.0",
     "remark-gfm": "^4.0.0",
     "styled-components": "^6.1.13"
   },
